+++ packages/ddp-client/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   summary: "Meteor's latency-compensated distributed data client",
-  version: '2.3.0-rc161.2',
+  version: '2.3.1-rc161.2',
   documentation: null
 });